feat: :rocket: refactor the code
