+++ src/pages/Home.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 import { useHistory } from 'react-router-dom';
 
 // import { auth, firebase } from '../services/firebase';
-import { AuthContext } from '../App';
+import { AuthContext } from '../contexts/AuthContext';
 
 import illustrationImg from '../assets/images/illustration.svg';
 import logoImg from '../assets/images/logo.svg';
