+++ src/pages/NewRoom.tsx
@@ -1,15 +1,16 @@
+// import { useContext } from 'react';
 import { Link } from 'react-router-dom'
 import { Button } from '../components/Button'
+// import { AuthContext } from '../contexts/AuthContext';
+
 import illustrationImg from '../assets/images/illustration.svg';
 import logoImg from '../assets/images/logo.svg';
 // import googleIconImg from '../assets/images/google-icon.svg';
 import '../styles/auth.scss'
-import { useContext } from 'react';
-import { AuthContext } from '../App';
 // webpack (Module Bundler) -> pega a extensão do arquivo e configura de forma predeterminada como cada um será lido no código
 
 export function NewRoom() {
-  const { user } = useContext(AuthContext);
+  // const { user } = useContext(AuthContext);
 
   return (
     <div id="page-auth">
@@ -21,7 +22,6 @@ export function NewRoom() {
       <main>
         <div className="main-content">
           <img src={logoImg} alt="Letmeask" />
-          <h1>{user?.name}</h1>
           <h2>Criar uma nova sala</h2>
           <form>
             <input
